Address comments from dominichamon for fr-1051
Committer: Hai Huang <haih@google.com>

On branch fr-1051
Your branch is up to date with 'origin/fr-1051'.

Changes to be committed:
modified:   src/benchmark.cc
modified:   src/benchmark_adjust_repetitions.cc
modified:   src/benchmark_adjust_repetitions.h
modified:   src/benchmark_api_internal.cc
modified:   src/benchmark_api_internal.h
modified:   test/benchmark_random_interleaving_gtest.cc

+++ src/benchmark_adjust_repetitions.h
@@ -28,12 +28,12 @@ struct InternalRandomInterleavingRepetitionsInput {
   double real_time_used_per_repetition;
   double min_time_per_repetition;
   double max_overhead;
-  size_t max_repetitions;
+  int max_repetitions;
 };
 
 // Should be called right after the first repetition is completed to estimate
 // the number of iterations.
-size_t ComputeRandomInterleavingRepetitions(
+int ComputeRandomInterleavingRepetitions(
     InternalRandomInterleavingRepetitionsInput input);
 
 }  // end namespace internal
